refactor(router): user api
