+++ src/router/src/apis/user/mod.rs
@@ -1,0 +1,1 @@
+pub mod api;
